+++ src/renderer/lib/monaco/language.ts
@@ -1,4 +1,5 @@
 import { languages } from 'monaco-editor';
+import xmlFormat from 'xml-formatter';
 import { TemplateVariableCompletionItemsProvider } from './template-variable-completion-items-provider';
 import { TemplateVariableHoverProvider } from './template-variable-hover-provider';
 import { TemplateVariableSemanticTokensProvider } from './template-variable-semantic-tokens-provider';
@@ -10,7 +11,7 @@ export enum Language {
 }
 
 const supportedLanguages = [Language.JSON, Language.XML, Language.TEXT];
-const formattableLanguages = [Language.JSON, Language.XML];
+const formattableLanguages = new Set([Language.JSON, Language.XML]);
 
 // plaintext using fast syntax highlighting
 languages.setMonarchTokensProvider(Language.TEXT, {
@@ -34,12 +35,26 @@ languages.registerCompletionItemProvider(
 // hover provider (shows template variable value on hover)
 languages.registerHoverProvider(supportedLanguages, new TemplateVariableHoverProvider());
 
+// XML formatting provider
+languages.registerDocumentFormattingEditProvider(Language.XML, {
+  provideDocumentFormattingEdits: (model) => {
+    const text = model.getValue();
+    const formattedText = xmlFormat(text);
+    return [
+      {
+        range: model.getFullModelRange(),
+        text: formattedText,
+      },
+    ];
+  },
+});
+
 /**
  * Determines whether the specified language supports code formatting.
  *
- * @param {string | Language} [language] - The language to check, either as a string or a `Language` enum value.
- * @returns {boolean} - Returns `true` if the language supports formatting; otherwise, `false`.
+ * @param language The language to check, as a string or enum value.
+ * @returns `true` if the language supports formatting; otherwise, `false`.
  */
 export const isFormattableLanguage = (language?: string | Language): boolean => {
-  return formattableLanguages.includes(language as Language);
+  return formattableLanguages.has(language as Language);
 };
